+++ backends/carp_backend/lib/carp_data_manager.dart
@@ -47,6 +47,10 @@ class CarpDataManager extends AbstractDataManager {
   @override
   String get type => DataEndPointTypes.CAWS;
 
+  /// Should data be compressed / zipped before upload?
+  bool get compress => carpEndPoint.compress;
+  set compress(bool compress) => carpEndPoint.compress = compress;
+
   /// The connectivity status of this data manager.
   List<ConnectivityResult> get connectivity => _connectivity;
   set connectivity(List<ConnectivityResult> status) {
@@ -134,15 +138,13 @@ class CarpDataManager extends AbstractDataManager {
           await CarpDataStreamService().appendToDataStreams(
             studyDeploymentId,
             batches,
-            compress: false,
+            compress: compress,
           );
           addEvent(
               DataManagerEvent(CarpDataManagerEventTypes.dataStreamAppended));
           break;
         case CarpUploadMethod.datapoint:
-          await uploadDataStreamBatchesAsDataPoint(
-            batches,
-          );
+          await uploadDataStreamBatchesAsDataPoint(batches);
           addEvent(DataManagerEvent(
               CarpDataManagerEventTypes.dataPointsBatchUploaded));
           break;