+++ packages/carp_movesense_package/lib/carp_movesense_package.dart
@@ -75,6 +75,7 @@ part 'movesense_device_manager.dart';
 class MovesenseSamplingPackage implements SamplingPackage {
   static const String MOVESENSE_NAMESPACE = "${NameSpace.CARP}.movesense";
 
+  static const String DEVICE_INFO = "$MOVESENSE_NAMESPACE.deviceinformation";
   static const String STATE = "$MOVESENSE_NAMESPACE.state";
   static const String HR = "$MOVESENSE_NAMESPACE.hr";
   static const String ECG = "$MOVESENSE_NAMESPACE.ecg";
@@ -95,6 +96,7 @@ class MovesenseSamplingPackage implements SamplingPackage {
 
   @override
   Probe? create(String type) => switch (type) {
+        DEVICE_INFO => MovesenseDeviceProbe(),
         STATE => MovesenseStateChangeProbe(),
         HR => MovesenseHRProbe(),
         ECG => MovesenseECGProbe(),
@@ -111,6 +113,7 @@ class MovesenseSamplingPackage implements SamplingPackage {
   void onRegister() {
     FromJsonFactory().registerAll([
       MovesenseDevice(),
+      MovesenseDeviceInformation(),
       MovesenseStateChange(MovesenseDeviceState.unknown),
       MovesenseHR(55),
       MovesenseECG(0, []),
@@ -122,6 +125,13 @@ class MovesenseSamplingPackage implements SamplingPackage {
   @override
   DataTypeSamplingSchemeMap get samplingSchemes =>
       DataTypeSamplingSchemeMap.from([
+        DataTypeSamplingScheme(
+          DataTypeMetaData(
+            type: DEVICE_INFO,
+            displayName: "Device Information",
+            timeType: DataTimeType.POINT,
+          ),
+        ),
         DataTypeSamplingScheme(
           DataTypeMetaData(
             type: STATE,