+++ packages/carp_movesense_package/lib/movesense_probes.dart
@@ -6,7 +6,7 @@
 
 part of 'carp_movesense_package.dart';
 
-abstract class _MovesenseProbe extends StreamProbe {
+abstract class _MovesenseStreamProbe extends StreamProbe {
   int? _subscriptionId;
 
   final StreamController<String> _streamController =
@@ -16,47 +16,59 @@ abstract class _MovesenseProbe extends StreamProbe {
   MovesenseDeviceManager get deviceManager =>
       super.deviceManager as MovesenseDeviceManager;
 
-  String get serial => deviceManager.serial;
-  String uri;
-  Function converter;
+  String get _serial => deviceManager.serial;
 
-  _MovesenseProbe(this.uri, this.converter);
+  final String _uri;
+  final Function _converter;
+
+  _MovesenseStreamProbe(this._uri, this._converter);
 
   @override
   Stream<Measurement>? get stream => deviceManager.isConnected
       ? _streamController.stream.map((event) =>
-          Measurement.fromData(converter.call(jsonDecode(event)) as Data))
+          Measurement.fromData(_converter.call(jsonDecode(event)) as Data))
       : null;
 
   @override
-  Future<bool> onStart() {
+  Future<bool> onStart() async {
     var completer = Completer<bool>();
 
     // fast out of already subscribed to this type of measurement
-    if (_subscriptionId != null) completer.complete(false);
+    if (_subscriptionId != null) return false;
 
     try {
-      _subscriptionId = Mds.subscribe("$serial/$uri", "{}", (data, status) {
+      _subscriptionId = Mds.subscribe("$_serial/$_uri", "{}",
+          // onSuccess
+          (data, status) {
         debug('$runtimeType - OnSuccess, data: $data, status: $status');
         completer.complete(super.onStart());
-      }, (error, status) {
-        warning('$runtimeType - OnError, error: $error, status: $status');
-        _streamController.addError(error);
+      },
+          // onError
+          (error, status) {
+        var errorMsg = '$runtimeType - Error, error: $error, status: $status';
+        warning(errorMsg);
+        _streamController.addError(errorMsg);
         _subscriptionId = null;
         completer.complete(false);
-      }, (data) {
-        debug('$runtimeType - OnNotification, data: $data');
+      },
+          // onNotification
+          (data) {
         _streamController.add(data);
-      }, (error, status) {
-        warning(
-            '$runtimeType - OnSubscriptionError, error: $error, status: $status');
-        _streamController.addError(error);
+      },
+          // onSubscriptionError
+          (error, status) {
+        var errorMsg =
+            '$runtimeType - Subscription Error, error: $error, status: $status';
+        warning(errorMsg);
+        _streamController.addError(errorMsg);
         _subscriptionId = null;
         completer.complete(false);
       });
     } catch (error) {
-      warning('$runtimeType - Error, error: $error');
-      _streamController.addError(error);
+      var errorMsg =
+          '$runtimeType - Error when trying to subscribe to device - serial: $_serial, uri: $_uri, error: $error';
+      warning(errorMsg);
+      _streamController.addError(errorMsg);
       _subscriptionId = null;
       completer.complete(false);
     }
@@ -67,7 +79,6 @@ abstract class _MovesenseProbe extends StreamProbe {
   @override
   Future<bool> onStop() async {
     super.onStop();
-    debug('$runtimeType - onStop,  _id: $_subscriptionId');
 
     if (_subscriptionId != null) Mds.unsubscribe(_subscriptionId!);
     _subscriptionId = null;
@@ -77,25 +88,25 @@ abstract class _MovesenseProbe extends StreamProbe {
 }
 
 /// A probe collecting [MovesenseHR] events.
-class MovesenseHRProbe extends _MovesenseProbe {
+class MovesenseHRProbe extends _MovesenseStreamProbe {
   MovesenseHRProbe() : super("Meas/HR", MovesenseHR.fromMovesenseData);
 }
 
 /// A probe collecting [MovesenseECG] events at 125 Hz.
-class MovesenseECGProbe extends _MovesenseProbe {
+class MovesenseECGProbe extends _MovesenseStreamProbe {
   MovesenseECGProbe() : super("Meas/ECG/125", MovesenseECG.fromMovesenseData);
 }
 
 /// A probe collecting [MovesenseTemperature] events.
 ///
 /// Note that not all type of Movesense devices supports temperature.
-class MovesenseTemperatureProbe extends _MovesenseProbe {
+class MovesenseTemperatureProbe extends _MovesenseStreamProbe {
   MovesenseTemperatureProbe()
       : super("Meas/Temp", MovesenseTemperature.fromMovesenseData);
 }
 
 /// A probe collecting [MovesenseIMU] events at 13 Hz (lowest).
-class MovesenseIMUProbe extends _MovesenseProbe {
+class MovesenseIMUProbe extends _MovesenseStreamProbe {
   MovesenseIMUProbe() : super("Meas/IMU9/13", MovesenseIMU.fromMovesenseData);
 }
 
@@ -110,16 +121,39 @@ class MovesenseIMUProbe extends _MovesenseProbe {
 /// events. This has been tested on the MD and HR2 devices.
 ///
 /// Therefore, this probe **only** listens to single tap events.
-class MovesenseStateChangeProbe extends _MovesenseProbe {
+class MovesenseStateChangeProbe extends _MovesenseStreamProbe {
   MovesenseStateChangeProbe()
       : super("System/States/4", MovesenseStateChange.fromMovesenseData);
+}
+
+/// A probe collecting [MovesenseDeviceInformation] from the connected
+/// Movesense device.
+class MovesenseDeviceProbe extends MeasurementProbe {
+  @override
+  Future<Measurement?> getMeasurement() async {
+    // fast out if not connected
+    if (!deviceManager.isConnected) return null;
+
+    var serial = (deviceManager as MovesenseDeviceManager).serial;
+    var completer = Completer<Measurement>();
+
+    Mds.get(Mds.createRequestUri(serial, "/Info"), "{}", ((info, statusCode) {
+      var data =
+          MovesenseDeviceInformation.fromMovesenseData(json.decode(info));
+      completer.complete(Measurement.fromData(data));
+    }), (error, statusCode) {
+      completer.completeError('$runtimeType - error: $error');
+    });
+
+    return completer.future;
+  }
 }
 
 // MULTI STATE CHANGE PROBE BELOW - with notes
 
 /// Enumeration of the type of state changes available on the Movesense device.
 /// See https://www.movesense.com/docs/esw/api_reference/#systemstates
-enum MovensenseStateChange {
+enum MovesenseState {
   movement,
   battery,
   connectors,
@@ -160,11 +194,11 @@ class MovesenseMultiStateChangeProbe extends StreamProbe {
 
   @override
   Future<bool> onStart() async {
-    _addStateSubscription(MovensenseStateChange.movement);
-    _addStateSubscription(MovensenseStateChange.connectors);
-    _addStateSubscription(MovensenseStateChange.doubleTap);
-    _addStateSubscription(MovensenseStateChange.tap);
-    _addStateSubscription(MovensenseStateChange.freeFall);
+    _addStateSubscription(MovesenseState.movement);
+    _addStateSubscription(MovesenseState.connectors);
+    _addStateSubscription(MovesenseState.doubleTap);
+    _addStateSubscription(MovesenseState.tap);
+    _addStateSubscription(MovesenseState.freeFall);
 
     return super.onStart();
   }
@@ -175,7 +209,7 @@ class MovesenseMultiStateChangeProbe extends StreamProbe {
           MovesenseStateChange.fromMovesenseData(jsonDecode(event))))
       : null;
 
-  void _addStateSubscription(MovensenseStateChange state) {
+  void _addStateSubscription(MovesenseState state) {
     int stateId = state.index;
 
     // fast out of already subscribed to this type of state change.